CUMULUS-2309 refactor tests to use shared helper
