+++ packages/api/tests/endpoints/test-granules.js
@@ -132,97 +132,6 @@ async function setupBucketsConfig() {
   return { internalBucket: systemBucket, publicBucket: buckets.public.name };
 }
 
-/**
- * Helper for creating a granule, and files belonging to that granule
- * @param {Knex} dbClient - Knex client
- * @param {number} collectionCumulusId - cumulus_id for the granule's parent collection
- * @param {boolean} published - if the granule should be marked published to CMR
- * @returns {Object} fake granule object
- */
-async function createGranuleAndFiles(dbClient, collectionCumulusId, published) {
-  const s3Buckets = {
-    protected: {
-      name: randomId('protected'),
-      type: 'protected',
-    },
-    public: {
-      name: randomId('public'),
-      type: 'public',
-    },
-  };
-  const granuleId = randomId('granule');
-  const files = [
-    {
-      bucket: s3Buckets.protected.name,
-      fileName: `${granuleId}.hdf`,
-      key: `${randomString(5)}/${granuleId}.hdf`,
-    },
-    {
-      bucket: s3Buckets.protected.name,
-      fileName: `${granuleId}.cmr.xml`,
-      key: `${randomString(5)}/${granuleId}.cmr.xml`,
-    },
-    {
-      bucket: s3Buckets.public.name,
-      fileName: `${granuleId}.jpg`,
-      key: `${randomString(5)}/${granuleId}.jpg`,
-    },
-  ];
-
-  const newGranule = fakeGranuleFactoryV2(
-    {
-      granuleId: granuleId,
-      status: 'failed',
-      published: published,
-    }
-  );
-
-  newGranule.files = files;
-
-  await createS3Buckets([
-    s3Buckets.protected.name,
-    s3Buckets.public.name,
-  ]);
-
-  // Add files to S3
-  await Promise.all(newGranule.files.map((file) => s3PutObject({
-    Bucket: file.bucket,
-    Key: file.key,
-    Body: `test data ${randomString()}`,
-  })));
-
-  // create a new Dynamo granule
-  await granuleModel.create(newGranule);
-
-  // create a new PG granule
-  const newPGGranule = fakeGranuleRecordFactory(
-    {
-      granule_id: granuleId,
-      status: 'failed',
-      collection_cumulus_id: collectionCumulusId,
-      published: published,
-    }
-  );
-
-  const [granuleCumulusId] = await granulePgModel.create(dbClient, newPGGranule);
-
-  // create PG files
-  await Promise.all(
-    files.map((f) => {
-      const pgFile = {
-        granule_cumulus_id: granuleCumulusId,
-        bucket: f.bucket,
-        file_name: f.fileName,
-        key: f.key,
-      };
-
-      return filePgModel.create(dbClient, pgFile);
-    })
-  );
-
-  return { s3Buckets, newGranule };
-}
-
 test.before(async (t) => {
   process.env.CMR_ENVIRONMENT = 'SIT';
   process.env = {
@@ -700,13 +609,13 @@ test('DELETE returns 404 if granule does not exist', async (t) => {
 });
 
 test('DELETE deleting an existing granule that is published will fail and not delete records', async (t) => {
-  const { s3Buckets, newGranule } = await createGranuleAndFiles(
+  const { s3Buckets, newPgGranule, files } = await createGranuleAndFiles(
     t.context.knex,
     t.context.collectionCumulusId,
     true
   );
 
-  const granuleId = newGranule.granuleId;
+  const granuleId = newPgGranule.granule_id;
 
   const response = await request(app)
     .delete(`/granules/${granuleId}`)
@@ -727,7 +636,7 @@ test('DELETE deleting an existing granule that is published will fail and not de
 
   // Verify files still exist in S3 and PG
   await Promise.all(
-    newGranule.files.map(async (file) => {
+    files.map(async (file) => {
       t.true(await s3ObjectExists({ Bucket: file.bucket, Key: file.key }));
       t.true(await filePgModel.exists(t.context.knex, { bucket: file.bucket, key: file.key }));
     })
